refactor(AcceptanceTest): 인수테스트 요청 API 분리
