+++ src/test/java/kitchenpos/table/TableGroupAcceptanceTest.java
@@ -1,18 +1,16 @@
 package kitchenpos.table;
 
+import static kitchenpos.table.TableAcceptanceAPI.단체_손님을_생성;
+import static kitchenpos.table.TableAcceptanceAPI.단체_손님을_해제;
+import static kitchenpos.table.TableAcceptanceAPI.손님_입장;
 import static kitchenpos.table.TableAcceptanceTest.빈자리;
-import static kitchenpos.table.TableAcceptanceTest.손님_입장;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
 import kitchenpos.AcceptanceTest;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.TableGroup;
-import kitchenpos.table.dto.TableGroupRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -20,7 +18,6 @@ import org.springframework.http.HttpStatus;
 
 class TableGroupAcceptanceTest extends AcceptanceTest {
 
-    TableGroup 단체손님;
     OrderTable 손님1;
     OrderTable 손님2;
 
@@ -53,20 +50,6 @@ class TableGroupAcceptanceTest extends AcceptanceTest {
 
         // then
         단체_손님이_해제됨(response);
-    }
-
-
-    public static ExtractableResponse<Response> 단체_손님을_생성(List<OrderTable> orderTableList) {
-        List<Long> orderTableIds = orderTableList.stream()
-                .map(OrderTable::getId)
-                .collect(Collectors.toList());
-        TableGroupRequest tableGroup = new TableGroupRequest(orderTableIds);
-
-        return AcceptanceTest.doPost("/api/table-groups", tableGroup);
-    }
-
-    public static ExtractableResponse<Response> 단체_손님을_해제(Long orderTableId) {
-        return AcceptanceTest.doDelete("/api/table-groups/" + orderTableId);
     }
 
     public static void 단체_손님이_생성됨(ExtractableResponse<Response> response) {
